Respond to PR comment: create JsonUtils class
